Fix compilation of wxNativeWindow sample for old compilers.
Define BS_SPLITBUTTON ourselves if it's not defined in the headers to at
least create the control even then. Skip the code handling BCN_DROPDOWN as
it's too much trouble to make it compile with the old headers however.

This should fix the build with VC8 currently failing on buildbot.