Revert "Optimize drawing of small bitmaps with alpha in wxMSW"
This reverts commit 4abf781f5443fdd85c8a045c6b3fccfd7e2ece51 because it
results in crashes due to writing out of bounds of the bitmap: nothing
guarantees that the entire (x, y, dstWidth, dstHeight) rectangle fits in the
destination bitmap and so the code of this commit could merrily overflow it
and did it as could be seen e.g. in the HTML test sample after scrolling
around a little.

See #16766.