Fix the build when using mismatching strict ANSI modes with MinGW
Don't use __STRICT_ANSI__ to determine whether a function is available or not
as its value can be different when building the library and the application,
resulting in mysterious link errors.

Instead use the same approach as in 02e79d24bda6c5e025a9336f03b300e5a199847c
and just declare the functions that are available but just not declared in
strict ANSI mode manually.

Closes #15793.