Really fix initial text selection in wxTextEntryDialog
Focus the text control after changing its value, this ensures that all text in
it is initially selected, allowing the user to easily replace it if necessary.

This was also the intention of b063c6bfc6a07c2cbecd569bb365d6548b48bef5, but
it was insufficient for wxGTK where changing the text control value seems to
reset its selection, unlike in wxMSW where the selection is preserved.

See https://github.com/wxWidgets/wxWidgets/pull/266