Fix wxMSW ListCtrl drawing of horizontal rules for new items
When adding a new item to a wxMSW ListCtrl that uses horizontal rules the
painting is horizontally clipped to the left of the control and the
rightmost column and the rule doesn't extend beyond that as it does when
doing a full repaint. Fix by refreshing the part to the right of the
rightmost column.

This also reverts efb0d15beec618c0d3eef19dddc96f562306a71e which became
unnecessary now.

Also see #17158.