Fix GetInstallPrefix availability check
Add wxHAS_STDPATHS_INSTALL_PREFIX to make checking for
wxStandardPaths::GetInstallPrefix()'s availability simpler, as the
condition under which it is compiled is nontrivial.

Fixes compilation after 7fb8b11a02b132c162198662fca3a67601c3392e on
Cygwin, which is a UNIX, but its wxStandardPaths implementation is a
Windows one.