Select the bitmap out of wxDC before using it in the unit test
The unit test added in 8997c95012d64cc3f73a46731005484e068d7e3e had a bug as
it tried to use the bitmap directly while it was still selected into a
wxMemoryDC, which wasn't guaranteed to work and resulted in an assert.

Fix this by destroying wxMemoryDC earlier.

See #17666.