Fix bug due to undefined g_wxAnyValueTypeGlobals initialization order
This fixes the changes of commit 84485aa05ce633dbea91a2e533e00b681b97b84e
which didn't work if any of the globals using wxPreRegisterAnyToVariant() were
instantiated before g_wxAnyValueTypeGlobals itself.

Wrap the global inside a function to ensure that we initialized it before it
is used by wxPreRegisterAnyToVariant() and not after.