Make wxBasicString as simple and safe to use as possible
Don't try to write a general purpose class, but provide just the methods
that we need in our code.

This fixes the bug added in a94b46bb4905beac143ee8b3d3788c646935d703
which resulted in a crash because a literal string, not a BSTR, was
passed to wxBasicString ctor.