Remove state assigning from wxProgressDialog::GetHandle()
This was added in 01bd848eb9699a60e0bdd5105b9a9c591c9e1b20 without any
explanation and probably was a copy-and-paste typo as it just doesn't
make any sense to change the state of the dialog in this accessor
function (and if the state doesn't change, then this assignment is just
completely useless).

Remove the apparently unnecessary assignment and also an unnecessary
temporary variable.