Fix recently introduced crash in ClippingBoxTestCase
Commit f8f002659db1a3ec7e06617de671c39688343635 did fix a memory leak in
this test case, but at the price of introducing a crash due to deleting
the same pointer twice.

The real fix would be to change the code here to avoid returning a
pointer which sometimes needs to be deleted and sometimes must not, but
for now just add a crude check to avoid crashing.