Try to fix macOS build using cmake after xlocale changes
Commit 6ce3c5dd4333b5bc18b2c818f4be9b48d6f54fe4 removed the inclusion of
xlocale.h because it is not (and never was) needed under Linux with
glibc, but it is still needed under macOS, so this (silently) disabled
wxXLocale support under Mac when using configure and broke the build
when using cmake.

Fix both problems by using xlocale.h only if it's available, both in
configure and in cmake.