Adjust wxAuiMDIClientWindow tabs height to icons automatically
Don't explicitly set the default icon size, this isn't really necessary
and just prevents things from working correctly if the child frames use
icons of a different (especially bigger) size.

Still try to use the icons of the standard size by default, but let
wxAuiNotebook adjust to bigger icons if this is what we have.