Remove useless border style check from wxVListBox::Create()
The comparison could never be true as wxDEFAULT is not one of
wxBORDER_XXX values and so is not included in wxBORDER_MASK and so was
useless -- and just results in -Wtautological-compare from gcc 8.

Simply remove this check, a047aff270905b377a81657a90f92eb3297157d0 added
it apparently to use themed border by default for wxVListBox, but this
was done in a better way in 28319afe559ccc061d15e5a5cb7208959f5c771e.