Fix bug with showing TLW in wxMSW when restoring its geometry
Since the changes of 5d15970588b5d5c83e1b529f26671329a38cf085, the
windows were shown when their geometry was restored as a side effect of
calling ::SetWindowPlacement(). This was unexpected and resulted in
flicker on startup, so fix this by explicitly passing SW_HIDE to
SetWindowPlacement() if the window is currently hidden and storing the
real show command inside wxTLW itself, where it will be used when it's
finally shown.