Fix recent regression in wxMSW wxTopLevelWindow::IsMaximized()
Recent ef3d5b3cd05966b278fa4ef665e2508437f3bace broke IsMaximized()
return value when the window was visible and had been previously
maximized by calling Maximize(), but wasn't actually maximized any
longer because it started to always return true when m_showCmd was set
to SW_MAXIMIZE.

Fix this by only using m_showCmd when the window is hidden, as it
shouldn't matter what it is when it's shown. Also simplify/optimize the
logic of IsMaximized() to use either ::IsZoomed() or m_showCmd depending
on whether the window is shown or not, but not both.

See https://github.com/wxWidgets/wxWidgets/pull/842

Closes #18163.