Fix TLW appearance in macOS 10.14's dark mode
Set toplevel window's background colour to the correct value of
NSColor.windowBackgroundColor (i.e. wxSYS_COLOUR_APPWORKSPACE). Without this,
the window's frame would be rendered slightly incorrectly, because of the way
the dark mode blends colors - even when no part of the window's interior, where
the background color would be used, is visible.