CMake: Fix monolithic build
Monolithic build regressed with commit
815d288c4fedba044a9863c883d6dd9f53526668 "Set wx-config base, gui and
built libraries".

The code queries each possible target name and creates lists of the
valid targets. In the monolithic build, none of the normal target names
exist, so the list is empty. The empty list is then passed to STRIP,
causing cmake to fail due to not having enough arguments.

By quoting the STRIP params we can pass an empty argument and thus
prevent the error of not-enough arguments.

See https://github.com/wxWidgets/wxWidgets/pull/1100