Preserve recent changes to wxSTC documentation
Recent commits 958f5eb0d0 and b31ff5a2ce fixed some typos is the
documentation for wxSTC. Unfortunatly these changes will be lost the
next time the wxSTC files are regenerated. This commit modifies
stc.interface.h.in and gen_docs.py to ensure that these changes are
preserved when the wxSTC files are regenerated.