Fix assert in wxWindowX11::SetBackgroundColour(wxColour())
Setting invalid colour is supposed to work and reset the background to
the default one, but asserted and then crashed due to triggering another
assert while trying to repaint the window while the assert dialog was
shown, in wxX11.

Fix this in the most expedient way by just resetting the background
colour to its initial value.

A better fix would be to leave it invalid but provide an accessor
returning the default value if the background colour is not defined.