Preserve mask when copying wxBitmapRefData to avoid crashes
This crash was introduced in 1530f0ccc2da051cc4da106977c3f609da837db1
and could happens inside wxBitmap::MSWBlendMaskWithAlpha() because the
code checked for a mask and alpha, but the mask could be lost after
AllocExclusive().

In practice, this happened e.g. when using 16-bit color as is the case
for Windows 7 Hyper-V VMs.

Closes https://github.com/wxWidgets/wxWidgets/pull/1695