Fix wrong wxGrid::RefreshBlock calling for an empty grid
There is no cells to select if the grid is empty (no rows or no columns).
So check for the columns count in SelectRow and check for the rows count in
SelectCol and call RefreshBlock only if it necessary.

Fix https://trac.wxwidgets.org/ticket/18659
Fix https://trac.wxwidgets.org/ticket/18660