Move OLE initialization to a module
Previously this was done in wxMSW-specific wxApp version, which meant
that initialization was done after the modules initialization and,
especially problematically, uninitialization was done before modules
uninitialization, which resulted in problems when doing anything using
OLE from the modules OnExit().

This commit doesn't solve these problems yet, because it leaves modules
initialization order undefined, but the upcoming commits using it will.