Disallow resizing rows by dragging cell corners if disabled
Even when resizing rows by dragging their edges was disabled via
DisableDragRowSize(), it was still possible to try to resize them when
the mouse was over the cell corner.

Fix this and remove the special case for the corners entirely, it's not
necessary as the existing cases cover this one already.

Just rearrange them in the right order to prefer column resizing, as
this seems to be a much more commonly used operation than row resizing.