Revert "Fix wxMSW ListCtrl drawing of horizontal rules for new items"
This reverts commit 612373fa2a (Fix wxMSW ListCtrl drawing of horizontal
rules for new items, 2016-05-04) and also commit 8a706ba658 (Avoid
infinite repaint loop in wxListCtrl with wxLC_HRULES, 2020-07-24) which
fixed one problem with the original change, but not all of them, as it
was still easily possible to get into an infinite repaint loop.

So just return to the simplest and not too bad, even if not optimal,
solution of refreshing everything after changing a column width when
using horizontal rules.

See #17158, #18850.

Closes #18927.