Fix spurious assert when dropping items is impossible in wxMSW
Use correct drop effect in DropCleanup helper added in 24f5e34612 (Fix
problem with dragged icon remaining on screen under MSW 10, 2020-01-11):
using the value of the input pdwEffect argument was wrong, as it could
be a combination of multiple DROPEFFECT_XXX flags, when we really need a
single one here.

This fixes a regression and the code should do exactly the same thing
now when OnDrop() does not throw as it had done before that commit.

See #18499.

Closes #18965.