Translation docs update: don't recommend 20-year-old gettext; use msginit
Warning about gettext versions earlier than 0.10.35 is pointless,
as even 0.10.35 is from 2001, and so old it isn't even available
on GNU mirrors anymore. We shall not recommend a specific version in
the first place.

The proper way to start a new translation is to use msginit to initialize
a .po file -- just copying the .pot leaves many fields to their default
values, which the translator would need to fill in manually (the
Plural-Forms: field in particular).