Generalize wxBookCtrlBase::m_bookctrl to wxWindow
The field is used for common placement and sizing logic in
notebook-like controls that have a "controller buddy" control. Its use
isn't wxControl-specific and only relies on wxWindow API, so the
restriction to wxControl-inherited types is arbitrary.

Change the type to the more general wxWindow* pointer.

In practical terms, this allows derived classes to use e.g.
wxNativeWindow as the controller buddy.