Show translation from coreutils message catalog string
Actually test using the coreutils catalog, as this was broken for many
years due to using a wrong catalog name and not adding /usr/share/locale
as a search prefix, but went unnoticed.

So add a menu item to show translation of a string in this catalog to
make testing that this works simpler.