Revert commits trying to fix wxWebRequest tests under Wine
This reverts commits 9eff1d61a7 (Install python3 and pip for httpbin,
2021-06-27) and 16f0698247 (Install winbind package required for using
wxWebRequest with Wine, 2021-06-28) as we skip wxWebRequest tests anyhow
under Wine because they don't work with its WinHTTP implementation, so
running httpbin is not necessary any longer and so neither is installing
Python.

These commits are still preserved in history and this commit itself
could be reverted later if Wine WinHTTP becomes functional enough to run
wxWebRequest tests successfully.